requirements.txt
cryptography
pyscard
Pillow
python-dotenv
channels
nfcpy
Django
djangorestframework
python-dateutil